+++ react/components/EXPERIMENTAL_Table/index.tsx
@@ -100,6 +100,8 @@ export const tablePropTypes = {
       width: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
       sortable: PropTypes.bool,
       cellRenderer: PropTypes.func,
+      extended: PropTypes.bool,
+      condensed: PropTypes.arrayOf(PropTypes.string),
     })
   ),
   items: PropTypes.arrayOf(PropTypes.object),
@@ -141,9 +143,8 @@ export type TableComposites = {
 
 export type Items = object[]
 
-export type CellData = {
-  cellData: unknown
-  rowData: unknown
+export type ReturnedData = {
+  data: unknown | object
   rowHeight: number
   currentDensity: Density
   motion: ReturnType<typeof useTableMotion>
@@ -154,7 +155,9 @@ export type Column = {
   title?: string | Element | Function
   width?: number | string
   sortable?: boolean
-  cellRenderer?: (cellData: CellData) => React.ReactNode
+  cellRenderer?: (data: ReturnedData) => React.ReactNode
+  extended?: boolean
+  condensed?: string[]
 }
 
 Table.Toolbar = Toolbar