+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [9.108.5] - 2020-02-03
+
+### Added
+
+- `condensed` and `extended` columns to `EXPERIMENTAL_Table`.
+
+### Removed
+
+- `EXPERIMENTAL_Table` line actions.
+
 ## [9.108.4] - 2020-02-03
 
 ### Added